+++ src/utils/crc32.test.ts
@@ -2,6 +2,6 @@ import { crc32 } from './crc32'
 
 describe('utils:crc32', () => {
   test('is deterministic', () => {
-    expect(crc32('SYNOR')).toMatchInlineSnapshot(`2266284534`)
+    expect(crc32('SYNOR')).toMatchInlineSnapshot(`-2028682762`)
   })
 })